requirements.txt
asgiref==3.3.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.1.3
django-allauth==0.45.0
django-cors-headers==3.9.0
django-filter==21.1
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==4.14.2
idna==3.3
jwcrypto==1.0
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.7
validate-docbr==1.8.2
wrapt==1.13.2
